Release Payment To Additional Sellers 	- In order's tab, if the members is a group has not created flybuy account or has created flybuy account but has been marked as 		unverified by admin, then displayed a pop up regarding send notification to...
Summary:
...those members in a group 	- Send proper message to those members of the group 		- Added template and proper message content for those users

changed the name of the path

when saving the per unit commission, converted the value into float

Releasing payment to additional sellers 	- Created transaction from initial seller to synapse node id 	- and then again created a transaction from escrow to the additional seller account

Reviewers: sarbada

Reviewed By: sarbada

Maniphest Tasks: T1655

Differential Revision: http://phabricator.jyaasa.com/D1670

+++ app/controllers/dashboard/groups_controller.rb
@@ -146,7 +146,7 @@ class Dashboard::GroupsController < ApplicationController
 	      if params["group"]["additional_seller_delete"].present?
 	      	params["group"]["additional_seller_delete"].each do |group_seller_id|
 	      		group_seller = GroupSeller.find_by_id(group_seller_id)
-	      		group_seller.destroy
+	      		group_seller.destroy if group_seller.present?
 	      	end
 	      end
 	    end
@@ -156,7 +156,7 @@ class Dashboard::GroupsController < ApplicationController
 
 	def destroy
     group = Group.find_by_id(params["id"])
-    group.destroy
+    group.destroy if group.present?
 		redirect_to dashboard_groups_path
   end
